DefectDetection/requirements.txt
opencv-python
numpy
pyyaml
tqdm
matplotlib
wandb
--extra-index-url https://download.pytorch.org/whl/cu124
torch
torchvision
